// src/app/shared/components/logo/logo.enums.ts
export enum ELogoVariant {
  BOOKMARK = 'BOOKMARK',
  BOOKMARK_ANIMATED = 'BOOKMARK_ANIMATED',
  BOOKMARK_S = 'BOOKMARK_S',
  BOOKMARK_S_MIN = 'BOOKMARK_S_MIN',
  BOOKMARK_ANIMATED_S = 'BOOKMARK_ANIMATED_S',
  BOOKMARK_OUTLINE = 'BOOKMARK_OUTLINE',
  BOOKMARK_OUTLINE_S = 'BOOKMARK_OUTLINE_S',
  BOOKMARK_OUTLINE_ANIMATED = 'BOOKMARK_OUTLINE_ANIMATED',
  BOOKMARK_OUTLINE_ANIMATED_S = 'BOOKMARK_OUTLINE_ANIMATED_S',
}

export enum ELogoLetterFill {
  TRANSPARENT = 'TRANSPARENT',
  WHITE = 'WHITE',
}